Decode DTB properties to native type and drop schema fixups
Originally with the YAML encoded DT there was little to no information
about a property's type. So all integer properties were decoded into an
2 dimensional matrix even if the property was a scalar. However, in
order to keep schemas simple, the schemas were written according to the
type and omitted this complexity. So the tools processed the schemas to
add this extra encoding. This works okay except in a few places like
if/then schemas where the tools don't know the property type in order to
do the correct fixup. With the DTB based decoding, we now have the
property type for every property and we can just decode the properties
to the correct type.

This only affects integer types as 'string' and 'string-array' are still
both encoded the same way (as an array). And 'phandle' is still encoded
as a matrix like 'phandle-array'.

Signed-off-by: Rob Herring <robh@kernel.org>