schemas: Split up PCI binding
The pci-bus.yaml binding is mostly for host bridges. However, there is a
need to define a binding schema for PCI-PCI bridges and PCI devices.
Much of the pci-bus.yaml applies to PCI-PCI bridges as well. Split the
binding into separate bus, device, host-bridge and PCI-PCI bridge
bindings.

A complication is that pci-bus.yaml is already widely referenced, so we
need to preserve its name and all the host bridge properties. Just
leaving pci-bus.yaml as the host bridge binding would be confusing.
Instead, move everything to the new binding files and make pci-bus.yaml
just a reference to pci-host-bridge.yaml.

Signed-off-by: Rob Herring <robh@kernel.org>