schema: Another take on checking additionalProperties/unevaluatedProperties
Complete schemas need to have additionalProperties or
unevaluatedProperties constraint in order to prevent undefined
properties. Attempts to check this with meta-schemas resulted in corner
cases which couldn't be fixed or required redundant
unevaluatedProperties. The problem is we need to know if a referenced
schema has the constraint already or not. So let's solve this with code
to look into $ref's to see if the ref constrains the properties or not.

Signed-off-by: Rob Herring <robh@kernel.org>