fixups: Fix handling of required "interrupts" without "properties" in schema
In subschemas with "{required: [ interrupts ]}", we fail to do the
"interrupts-extended" fixup because there is not any "properties" entry.
Therefore, we need to do the fixup for "required" regardless unless it
is under a "oneOf" already (otherwise, we'd keep recursing applying the
same fixup).

Also, if "required" becomes empty, remove it.

Signed-off-by: Rob Herring <robh@kernel.org>