validator: Stop merging scalar types to matrix types
There's a problem with properties which are defined as a single value
and matrix/phandle-array. The matrix dimensions may not cover the single
value case. In order to handle this, record the differing type rather
than trying to merge the types. Then when decoding the property, we can
make a better decision on the type. For string and string-array, they
can still be merged.

May have to revisit this when we drop making everything a matrix.

Signed-off-by: Rob Herring <robh@kernel.org>