Revert "schemas: i2c-controller: Add 'no-detect' property"
This reverts commit ba55f96c6d8d8c58f375393bca0c251574986c40.

This property was accepted, but then users of it were rejected[1][2].
Wolfram Sang pointed this out:
"no-detect" is broken. At the least, it should be named something like
"bus-fully-described" and then the OS can decide to leave out
auto-detection mechanisms. If you are interested in the latter, you can
simply disable class based instantiation on the host controller. No
need to describe this in DT.

[1] https://lore.kernel.org/all/20220405121627.1560949-2-vincent.whitchurch@axis.com/
[2] https://lore.kernel.org/all/20220429130749.3032462-1-vincent.whitchurch@axis.com/

Signed-off-by: Rob Herring <robh@kernel.org>