schemas: dma: Ensure 'dma-names' entries are unique (mostly)
'dma-names' has to be treated special as there's some Renesas bindings
which abuse dmas/dma-names to have multiple providers for the same
thing. IOW, the platform has 2 DMA controllers and a device can use
either one. This should have probably been handled with some form of
nexus node instead.

In order to prevent this in other cases, only allow 'tx' and 'rx' to be
repeated.

Signed-off-by: Rob Herring <robh@kernel.org>