dtb: Filter out types with the wrong length multiple
If we have a property which can be a string or non-string type and its
value has empty strings in it, it gets incorrectly decoded a a non-string.
There's not a great way to solve this heuristic. Perhaps we could count
the number of non-zero length strings and pick string type if it is above
some percentage.

For now, when the length is not a multiple that fits the integer types, we
can for sure remove those integer types. If a string property happens to
have a length matching a integer multiple, then there's still a problem.

Signed-off-by: Rob Herring <robh@kernel.org>