validator: Rework selecting schemas for validation
Validation of devicetrees is slow because every schema is checked
whether to apply it or not to every node. With 1000s of schemas and
100s-1000s of nodes in 1000s of DTBs, that makes for long validation
times. The reality is there's a few schemas that apply to all nodes
(mostly the core schemas) and only 1 schema with a matching compatible
string that applies (not counting schemas included by $ref). There's
also a few corner cases which have their own 'select'.

Rework the validation to take advantage of this by storing up front the
list of schemas to always iterate thru, and a map of compatible strings
to their respective schema. The result is a 6-7x reduction in the
validation time.

Signed-off-by: Rob Herring <robh@kernel.org>