validator: Add a version to the processed schemas
The contents of the processed schemas can change either from fixes to
the fixups or new features. Add a version tag to the schema so we can
detect changes and refuse to use old processed schema file.

Really, there's enough information (the file paths) in the processed
schema that we could just re-run the processing, but that can be a
future enhancement.

Since we now require a version, there's no need to handle the old list
format of the processed schemas and that fixup can be dropped.

Signed-off-by: Rob Herring <robh@kernel.org>