dt-validate: Add an 'only matching compatible' schema validation mode
Add an 'only matching compatible' schema validation mode which only
applies the schema matching nodes' compatible string. Currently, only
the most specific (i.e. 1st) compatible string in the instance is used,
but that may change in the future. This mode tends to be 4x faster (on
top of the recent 6-7x speed up) while still giving 80-90% of the
warnings. Mostly the core schemas are skipped.

Signed-off-by: Rob Herring <robh@kernel.org>