Improve 'oneOf' error messages
'oneOf' failures are hard because the messages give little indication of
which of the oneOf clauses is most likely what needs to be fixed.

The jsonschema.best_match() method can help in some cases when the errors
have different schema depths.

The handling for more than 1 oneOf condition being true is worse than the
default for jsonschema as it prints the whole oneOf schema rather than just
the true cases. I think jsonschema improved this at some point. With that,
we can just remove the custom formatting.

Signed-off-by: Rob Herring (Arm) <robh@kernel.org>