schemas: simple-bus: Rework handling of child nodes without unit-address
The simple-bus currently will spit out the whole bus subtree when it
contains a child node without a unit-address. It also has a false positive
when the child node is another bus that has an empty 'ranges' property.

Rework the schema to use 'additionalProperties' instead of a pattern and
use an if/then schema. Now validation will report "ranges is required"
which is a little misleading, but better than the huge dump of the whole
subtree.

Signed-off-by: Rob Herring (Arm) <robh@kernel.org>