dtb: Ensure there's only 1 phandle type
Dropping the 'phandle-array' when the length is 4 bytes was wrong as having
a single phandle can still be valid in that case. Instead, make sure there
is only a single phandle type. Otherwise, we'll end up decoding as a
'uint32-array' which is the default when the type can't be determined.

Signed-off-by: Rob Herring (Arm) <robh@kernel.org>