dtb: Handle multiple types for 'dma-masters'
The 'dma-masters' property is problematic to decode as it can be either a
phandle-array or uint32. This didn't matter until commit 23441a412f88
("Decode DTB properties to native type and drop schema fixups") which
made decoding of these 2 types different. Ideally, we'd get rid of the
uint32 user, but for now just look at the value as the uint32 case can
only be 1-4 and a phandle is most likely (but still possible) not 1-4.

Signed-off-by: Rob Herring (Arm) <robh@kernel.org>