validator: Add warning for properties with incompatible multiple types
While generally having multiple types for the same property name is
discouraged, there are some cases of this we're stuck with. In many cases
the type for a specific instance can be determined based on the property
length or the data (ASCII strings). In a few cases, there is no way to
tell. These cases have types of scalar, array, or matrix, and the possible
property length overlaps.

Signed-off-by: Rob Herring (Arm) <robh@kernel.org>