schemas: root-node: Allow extending root node schema
The root node schema doesn't prevent unknown properties and yet at the
same time can't be referenced to be extended. Some platforms have custom
properties in the root node (though that is discouraged). Reowrk the
schema so it can be extended.

In testing this, standard properties 'interrupt-parent' and
'dma-coherent' were found in use, so add them here.

Signed-off-by: Rob Herring (Arm) <robh@kernel.org>