dtb: Fix handling of nodes with 'interrupts' and 'interrupt-map'
The incorrect number of interrupt cells was used for a node containing
'#interrupt-cells', 'interrupts', and 'interrupt-map'. The issue is
'interrupts' needs to be parsed with the parent '#interrupt-cells' while
'interrupt-map' needs to be parsed with the current node's
'#interrupt-cells'. The code was overwriting the former with the latter
while looping thru the node properties. Fix this by storing the
interrupt-map '#interrupt-cells' value in a separate variable.

Signed-off-by: Rob Herring (Arm) <robh@kernel.org>