schema: Rework $ref checks
The $ref checks are skipping a number of instances of $ref due to the
additionalProperties/unevaluatedProperties checks. The $ref checking needs
to come before the additionalProperties/unevaluatedProperties checks.

The check for false properties was only needed for if/then/else schemas
which we skip now, so it can be dropped.

Signed-off-by: Rob Herring (Arm) <robh@kernel.org>