schema: Rework instantiating DTSchema.validator
DTSchema.validator is lazily instantiated when iter_errors() is called to
avoid overhead when not validating. However, this means check_schema_refs()
fails if iter_errors() has not been called first. Move the instantiation
to a method instead. Also, the resolver is stored in the validator instance,
so we don't need to store it.

Signed-off-by: Rob Herring (Arm) <robh@kernel.org>