validator: Fix ordering issue with properties referencing another property
The logic to handle properties which are a reference to another property is
dependent on the order the properties are processed. If the referenced
property has not been processed, then the property with the reference won't
get a type. Fix this by saving the type as the referenced property name.
Then resolve the reference when all the properties have been processed.

Signed-off-by: Rob Herring (Arm) <robh@kernel.org>