dtb: Improve 'dma-masters' property type handling
In cases where the dma-masters phandle value happened to be 1-4, the
decoding to an uint32 was incorrect. Default to using phandle-array type
and move the checks later when we can check if the phandle points to a
valid DMA provider node. If it doesn't, it's a safe bet that uint32 is
the right type.

Signed-off-by: Rob Herring (Arm) <robh@kernel.org>