dts_style: Add basic DTS coding style checker
Add basic coding style checker for DTS files, following Linux kernel DTS
coding style.  The checker does not test syntax validity and whether the
code compiles in the first place.  That's the job for dtc.  Instead the
tool operates on DTS line-by-line, similarly to what Linux kernel
checkpatch.pl does.

The goal of the tool is to improve DTS readability, thus the source is
parsed instead of flattened form.  For that reason none of standard
libraries, like libfdt or dtc, are suitable.

Supported checks so far: node name naming.

Signed-off-by: Krzysztof Kozlowski <krzk@kernel.org>